Docker/fastapi/requirements.txt
fastapi
uvicorn
pycaret
pandas
numpy
pydantic
unidecode
requests
streamlit
PyJWT
datetime
